backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.3
Pillow>=10.4
razorpay>=1.4
celery>=5.4
django-celery-beat>=2.6
redis>=5.0
python-decouple>=3.8
